Category refactor: moved components into webapp/components
