+++ webapp/components/survey/CategoryDetails/CategoryDetails.scss
