+++ webapp/components/survey/CategoryDetails/ImportSummary/ImportSummary.scss
