+++ webapp/components/survey/CategoryDetails/ImportSummary/index.js
@@ -1,0 +1,1 @@
+export { default } from './ImportSummary'
