+++ webapp/views/App/views/Designer/Designer.js
@@ -9,8 +9,7 @@ import TaxonomyList from '@webapp/components/survey/TaxonomyList'
 import TaxonomyDetails from '@webapp/components/survey/TaxonomyDetails'
 
 import CategoryList from '@webapp/views/App/views/Designer/CategoryList'
-
-import CategoryView from '@webapp/loggedin/surveyViews/category/categoryView'
+import CategoryDetails from '@webapp/components/survey/CategoryDetails'
 
 import { appModules, appModuleUri, designerModules } from '@webapp/app/appModules'
 import { useAuthCanEditSurvey } from '@webapp/store/user'
@@ -60,7 +59,7 @@ const Designer = () => {
           },
 
           {
-            component: CategoryView,
+            component: CategoryDetails,
             path: `${appModuleUri(designerModules.category)}:categoryUuid`,
           },
